Add Guava 17 to system package as well (and intro. new PackagesBuilder)
Also (unrelated) new installBootBundles() method for proper ordering.

Also removes CommandRegistrationImpl which used ImmutableList from Guava
and made this problem appear in the first place.  Now that this is
removed it actually wouldn't be a problem anymore in the short term.

But because the Sponge API depends on Guava, it seems like a good idea
to have Guava 17 at the same classloader level nevertheless - just to be
safe.

This may change in the future when I make the Sponge API a bundle.